Added Appointment and exceptions

+++ src/main/java/com/example/DoctorAppointmentApp/dto/doctor/DoctorDTO.java
@@ -1,12 +1,13 @@
 package com.example.DoctorAppointmentApp.dto.doctor;
 
 import com.example.DoctorAppointmentApp.model.doctors.DoctorSpeciality;
-import lombok.*;
+import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 @Setter
 public class DoctorDTO {
-    private int id;
+    private long id;
     private String firstName;
     private String lastName;
     private DoctorSpeciality doctorSpeciality;
